app/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
codetiming==1.3.0
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
multidict==5.1.0
mysql-connector-python==8.0.24
numpy==1.20.3
pandas==1.2.4
Pillow==8.3.1
plotly==4.14.3
protobuf==3.16.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.16.0
SQLAlchemy==1.4.14
SQLAlchemy-Utils==0.37.4
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==1.0.1
yarl==1.6.3
